update battery capacity value type

+++ src/main/java/com/musala/dronemanagerservice/model/dto/BatteryDto.java
@@ -1,4 +1,4 @@
 package com.musala.dronemanagerservice.model.dto;
 
-public record BatteryDto(String serialNumber, Float batteryCapacity) {
+public record BatteryDto(String serialNumber, Byte batteryCapacity) {
 }
